Move some types from input.d.ts -> optuna.ts
