+++ optuna_dashboard/ts/components/Artifact/ThreejsArtifactViewer.tsx
@@ -9,7 +9,7 @@ import { PerspectiveCamera } from "three"
 import { Modal, Box, useTheme } from "@mui/material"
 import ClearIcon from "@mui/icons-material/Clear"
 import IconButton from "@mui/material/IconButton"
-import { Artifact } from "ts/types"
+import { Artifact } from "ts/types/optuna"
 
 export const isThreejsArtifact = (artifact: Artifact): boolean => {
   return (
