+++ optuna_dashboard/ts/components/BestTrialsCard.tsx
@@ -14,7 +14,7 @@ import {
 } from "@mui/material"
 import { Link } from "react-router-dom"
 import LinkIcon from "@mui/icons-material/Link"
-import { StudyDetail, Trial, URL_PREFIX } from "ts/types"
+import { StudyDetail, Trial } from "ts/types/optuna"
 
 const useBestTrials = (studyDetail: StudyDetail | null): Trial[] => {
   return useMemo(() => studyDetail?.best_trials || [], [studyDetail])
