+++ optuna_dashboard/ts/components/GraphEdf.tsx
@@ -4,7 +4,7 @@ import { Typography, useTheme, Box } from "@mui/material"
 import { Target, useFilteredTrialsFromStudies } from "../trialFilter"
 import { getCompareStudiesPlotAPI, CompareStudiesPlotType } from "../apiClient"
 import { usePlotlyColorTheme, useBackendRender } from "../state"
-import { StudyDetail, Trial } from "ts/types"
+import { StudyDetail, Trial } from "ts/types/optuna"
 
 const getPlotDomId = (objectiveId: number) => `graph-edf-${objectiveId}`
