+++ optuna_dashboard/ts/components/GraphParetoFront.tsx
@@ -18,7 +18,7 @@ import { PlotType } from "../apiClient"
 import { useBackendRender } from "../state"
 import { usePlot } from "../hooks/usePlot"
 import * as Optuna from "@optuna/types"
-import { StudyDetail, Trial, URL_PREFIX } from "ts/types"
+import { StudyDetail, Trial } from "ts/types/optuna"
 
 const plotDomId = "graph-pareto-front"
