+++ optuna_dashboard/ts/components/GraphSlice.tsx
@@ -22,7 +22,7 @@ import { useMergedUnionSearchSpace } from "../searchSpace"
 import { PlotType } from "../apiClient"
 import { usePlotlyColorTheme, useBackendRender } from "../state"
 import { usePlot } from "../hooks/usePlot"
-import { SearchSpaceItem, StudyDetail, Trial } from "ts/types"
+import { SearchSpaceItem, StudyDetail, Trial } from "ts/types/optuna"
 
 const plotDomId = "graph-slice"
