+++ optuna_dashboard/ts/components/GraphTimeline.tsx
@@ -7,7 +7,7 @@ import { PlotType } from "../apiClient"
 import { useBackendRender } from "../state"
 import { usePlot } from "../hooks/usePlot"
 import * as Optuna from "@optuna/types"
-import { StudyDetail, Trial } from "ts/types"
+import { StudyDetail, Trial } from "ts/types/optuna"
 
 const plotDomId = "graph-timeline"
 const maxBars = 100
