+++ optuna_dashboard/ts/hooks/useParamImportance.ts
@@ -3,7 +3,7 @@ import { useSnackbar } from "notistack"
 import { getParamImportances } from "../apiClient"
 import { useQuery } from "@tanstack/react-query"
 import { AxiosError } from "axios"
-import { ParamImportance } from "ts/types"
+import { ParamImportance } from "ts/types/optuna"
 
 export const useParamImportance = ({
   numCompletedTrials,
